requirements.txt
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0
rank-bm25==0.2.2
pymupdf==1.23.8
python-docx==1.1.0
beautifulsoup4==4.12.2
groq==0.4.2
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.3
reportlab==4.0.8